googlecode issue 26
Don't use the static class initialised singleton pattern, as classes are not reloaded in the eclipse plugin. Instead lazily create, but reuse the analysis. This fix assumes a single threaded model within the eclipse plugin, and that detectors are registered only once during each execution (to get the benefits of caching).

git-svn-id: http://mutability-detector.googlecode.com/svn/trunk/MutabilityDetector4FindBugs/trunk@564 c7a0535c-eda6-11de-83d8-6d5adf01d787